Get rid of the delay hack
When inside the trainer, the newly found exercises are now inserted
right after the current one. The exercises that already are within a
lesson are not touched.

How it used to be: when the user wants to look for new exercises, all
available exercises are delayed, then the found exercises are added to
the end of the lesson. They get immediate access to these exercises, but
then they have to "refresh" all the old exercises, even the ones they
personally haven't skipped.